fixed a problem with uniform spacing of ribs when creating a new limb or body.

+++ PAM/Classes/QuadMeshProject/PAMUtilities.h
@@ -15,10 +15,10 @@
 
 +(void)centroids:(std::vector<GLKVector2>&)centroids ribWidth:(std::vector<float>&)ribWidth
                                        forTwoFingerTouchPoint:(std::vector<GLKVector2>)touchPointsWorld
-                                         withNextCentroidStep:(float)step;
+                                         withNextCentroidStep:(double)step;
 
 +(void)centroids:(std::vector<GLKVector2>&)centroids forOneFingerTouchPoint:(std::vector<GLKVector2>)touchPointsWorld
-                                                       withNextCentroidStep:(float)step;
+                                                       withNextCentroidStep:(double)step;
 
 +(void)normals:(std::vector<GLKVector2>&)normals
       tangents:(std::vector<GLKVector2>&)tangents
